+++ opendit/models/latte.py
@@ -7,11 +7,11 @@ import torch.nn as nn
 import torch.utils.checkpoint
 from einops import rearrange, repeat
 
-from opendit.dit.modules import DiTBlock, FinalLayer
 from opendit.embed.clip_text_emb import TextEmbedder
 from opendit.embed.patch_emb import PatchEmbed3D
 from opendit.embed.pos_emb import get_1d_sincos_pos_embed, get_2d_sincos_pos_embed
 from opendit.embed.time_emb import TimestepEmbedder
+from opendit.modules.block import DiTBlock, FinalLayer
 
 
 class Latte(nn.Module):
